+++ include/iutest_ver.hpp
@@ -17,11 +17,11 @@
 
 //======================================================================
 // define
-#define IUTEST_VER          0x01170009u     //!< iutest version 1.17.0.9
+#define IUTEST_VER          0x01170011u     //!< iutest version 1.17.0.11
 #define IUTEST_MAJORVER     0x01u           //!< Major Version
 #define IUTEST_MINORVER     0x17u           //!< Minor Version
 #define IUTEST_MICROVER     0x00u           //!< Micro Version
-#define IUTEST_REVISION     0x09u           //!< Revision
+#define IUTEST_REVISION     0x11u           //!< Revision
 
 #define IUTEST_BUILD        IUTEST_MICROVER //!< @deprecated
 
@@ -124,11 +124,14 @@ modification, are permitted provided that the following conditions are met: \n
   <li>v1.17.1.0
     <ul>
       <li>GTEST_SKIP 対応</li>
+      <li>IUTEST_*_NE_RANGE アサーションを追加</li>
+      <li>IUTEST_*_NE_COLLECTIONS アサーションを追加</li>
       <li>::std::wstring_view / ::std::u16string_view / ::std::u32string_view 対応</li>
       <li>アサーションに operator &lt;&lt; したときの出力を PrintToString したものと同じになるように変更</li>
       <li>テストをリピートした際のレポート出力ファイル名の書式を printf 書式から {I} に変更({I} にリピート回数が置換されます)</li>
       <li>Python2 のサポート終了</li>
       <li>C++20 で削除された basic_ostream::operator &lt;&lt; に対応</li>
+      <li>IUTEST_AP が同じ名前空間に複数定義できない問題を修正</li>
       <li>Google Test 旧バージョンとの互換性を修正</li>
     </ul>
   </li>