+++ CHANGES.md
@@ -6,6 +6,8 @@
 
 * New
   * GTEST_SKIP 対応
+  * IUTEST_*_NE_RANGE アサーションを追加
+  * IUTEST_*_NE_COLLECTIONS アサーションを追加
 
 * Improved
   * ::std::wstring_view / ::std::u16string_view / ::std::u32string_view 対応
@@ -17,6 +19,8 @@
 
 * Bug fixes
   * C++20 で削除された basic_ostream::operator << に対応
+  * IUTEST_AP が同じ名前空間に複数定義できない問題を修正
+  * Google Test 旧バージョンとの互換性を修正
   * いくつかの不具合を修正
 
 --------------------------------------------------